Separated Location request and HTTP handler
